move ClusterServiceStatus to types.ts
