+++ app/reducers/managerReducer/clusterStatusReducer.ts
@@ -1,7 +1,6 @@
 import type { Reducer } from 'redux';
 import * as types from '../../actions/types';
-import type { ClusterServiceStatus } from '../../components/shared/cluster/consts';
-import type { ClusterServices } from '../../components/shared/cluster/types';
+import type { ClusterServices, ClusterServiceStatus } from '../../components/shared/cluster/types';
 
 export interface ClusterStatusData {
     isFetching?: boolean;
